Commit files for testing out Laravel
I was following the tutorials made by net ninja about the laravel
framework up to the video 14(more on migrations), which I honestly
don't have any use case but I want to familiarize myself with laravel
even if the content is a out of my reach and understanding.

All the shit is confusing. But looks cool as fuck.

This commit is up to the part where I tried to run the
`php artisan migrate` command which is about creating tables inside
the database registered in the `.env` file in the project directory.

I don't know why I have to uncomment the `extension:pdo_mysql` in the
php.ini file. I think I should not have done that part.

But all is worth it since the command got executed fine after changing
the `php.ini` file. Followed the steps indicated in the answers for
similar problems in stack overflow.