Working DB Table Content Display
All the database tables can be displayed properly within a table.

A fucking pain to implement. Study Eloquent ORM- Seriously- jumping
straight to a framework without figuring out the basics of OOP in PHP
is a bad decision.

I need to understand the method signatures and the return types of the
method calls of the Eloquent ORM. Code will likely change due to bugs.

There are a lot of code to test out the Controllers and View, might
need to remove or edit them soon.