Optimized routes & added Screenshots
I changed the routes to fit the need of the project. I did not
include the original `welcome` view since that has no connection to the
project at all. The design of the project is shit but it is of no
high priority for now.

The Screenshots about the project will be added to the repository for
other members to see.